refactored naming of get profile
