+++ social-media-angular/src/app/app.module.ts
@@ -15,7 +15,8 @@ import { CommentComponent } from './components/comment/comment.component';
 import { UserCardComponent } from './components/user-card/user-card.component';
 import { UserInitialsPipe } from './pipes/user-initials.pipe';
 import { NavbarComponent } from './components/navbar/navbar.component';
-import { GetProfileComponent } from './components/get-profile/get-profile.component';
+import { GetProfileComponent } from './components/profile-get/get-profile.component';
+
 
 @NgModule({
   declarations: [
@@ -29,6 +30,7 @@ import { GetProfileComponent } from './components/get-profile/get-profile.compon
     NavbarComponent,
     UserInitialsPipe,
     GetProfileComponent,
+   
   ],
   imports: [
     BrowserModule,
