+++ social-media-angular/src/app/components/post-feed-page/post-feed-page.component.html
@@ -34,3 +34,4 @@
         <app-post [post]="post"></app-post>
     </div>
 </div>
+
