+++ social-media-angular/src/app/components/post-feed-page/post-feed-page.component.ts
@@ -4,6 +4,7 @@ import Post from 'src/app/models/Post';
 import User from 'src/app/models/User';
 import { AuthService } from 'src/app/services/auth.service';
 import { PostService } from 'src/app/services/post.service';
+import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-post-feed-page',
@@ -21,7 +22,8 @@ export class PostFeedPageComponent implements OnInit {
   posts: Post[] = [];
   createPost:boolean = false;
 
-  constructor(private postService: PostService, private authService: AuthService) { }
+  constructor(private postService: PostService, private authService: AuthService,
+    private router : Router) { }
 
   ngOnInit(): void {
     this.postService.getAllPosts().subscribe(
@@ -45,4 +47,6 @@ export class PostFeedPageComponent implements OnInit {
         }
       )
   }
+
+  
 }
